fix: move schemas to models files GardenPlan and PlantPlacement
